scripts/requirements.txt
PyGithub==2.2.0
python-slugify==8.0.1
scikit-learn>=1.0.0
joblib>=1.0.0
